Added tool to scam pagerank by method 1
This method will suggest a set of pages that you can edit to add a link
to the page you want to boost, in order to boost your normalized
pagerank by a certain amount. We also have a function that will run your
proposed set of changes and see how much the pagerank actually changes
by. It's in scam_pagerank.py